Move certain lines of code to makeGrid func
Had to move a couple lines to the makeGrid function, specifically the lines
that were adding the mouseover event listener to each cell. It was not running
that loop when the graph was reset, thus not allowing mouseovers to change the
classes of the cells on any graph created after the first one.